requirements.txt
Flask>=2.0
PyYAML
bcrypt
requests

